requirements.txt
filterpy
numpy
opencv-python
Pillow
scipy
PyYAML
torch>=1.10.0
torchvision>=0.11.1
tqdm
timm>=0.9.0
